OY-3920 : HakukohdeActionBar siirretty etusivulle
