+++ src/main/app/src/pages/HomePage/HakukohteetSection/index.tsx
@@ -1,19 +1,18 @@
-import React, { useMemo } from 'react';
+import React, { useCallback, useMemo } from 'react';
 
+import _fp from 'lodash/fp';
 import { useTranslation } from 'react-i18next';
 
 import Button from '#/src/components/Button';
-import {
-  makeHakuColumn,
-  makeKoulutustyyppiColumn,
-  makeModifiedColumn,
-  makeMuokkaajaColumn,
-  makeNimiColumn,
-  makeTilaColumn,
-} from '#/src/components/ListTable';
 import { ENTITY, ICONS } from '#/src/constants';
 import useModal from '#/src/hooks/useModal';
 import { hakukohdeService } from '#/src/machines/filterMachines';
+import { EntityListActionBar } from '#/src/pages/HomePage/EntityListActionBar';
+import { createHakukohdeListColumns } from '#/src/pages/HomePage/HakukohteetSection/createHakukohdeListColumns';
+import {
+  SERVICE_BY_ENTITY,
+  useEntitySelection,
+} from '#/src/pages/HomePage/useEntitySelection';
 import { useFilterState } from '#/src/pages/HomePage/useFilterState';
 import { searchHakukohteet } from '#/src/utils/hakukohde/searchHakukohteet';
 
@@ -24,24 +23,34 @@ import LiitoksetModal from './LiitoksetModal';
 
 const { HAKUKOHDE } = ENTITY;
 
-const useTableColumns = (t, organisaatioOid) =>
-  useMemo(
-    () => [
-      makeNimiColumn(t, {
-        getLinkUrl: ({ oid }) =>
-          `/organisaatio/${organisaatioOid}/hakukohde/${oid}/muokkaus`,
-      }),
-      makeHakuColumn(t, {
-        getLinkUrl: ({ hakuOid }) =>
-          `/organisaatio/${organisaatioOid}/haku/${hakuOid}/muokkaus/`,
-      }),
-      makeKoulutustyyppiColumn(t),
-      makeTilaColumn(t),
-      makeModifiedColumn(t),
-      makeMuokkaajaColumn(t),
-    ],
-    [t, organisaatioOid]
+const HakukohdeActionBar = () => {
+  const { selection, removeSelection } = useEntitySelection(HAKUKOHDE);
+
+  const changeTila = useCallback(
+    tila => {
+      //TODO seivataan tallennuksen yhteydessä.
+      if (tila !== null && !_fp.isEmpty(selection)) {
+        console.log('testi tila');
+        console.log(tila);
+        console.log('testi selection');
+        _fp.forEach(hakukohde => {
+          console.log(hakukohde);
+          // hakukohde.tila = tila?.value;
+        }, selection);
+      }
+    },
+    [selection]
   );
+  return (
+    <EntityListActionBar
+      entityType={HAKUKOHDE}
+      selection={selection}
+      removeSelection={removeSelection}
+      changeTila={changeTila}
+      copyEntities={undefined}
+    />
+  );
+};
 
 const Actions = ({ organisaatioOid }) => {
   const { isOpen, close, open } = useModal();
@@ -62,9 +71,22 @@ const Actions = ({ organisaatioOid }) => {
 const HakukohteetSection = ({ organisaatioOid, canCreate = true }) => {
   const { t } = useTranslation();
 
+  // const isYhteishaku = isYhteishakuHakutapa(haku?.hakutapaKoodiUri);
+
+  const columns = useMemo(
+    () =>
+      createHakukohdeListColumns(
+        t,
+        organisaatioOid
+      )(SERVICE_BY_ENTITY[HAKUKOHDE]),
+    [t, organisaatioOid]
+  );
+
+  // const actionBar = isYhteishaku ? HakukohdeActionBar : undefined;
+
   const filterState = useFilterState(HAKUKOHDE, hakukohdeService);
 
-  const columns = useTableColumns(t, organisaatioOid);
+  // const columns = useTableColumns(t, organisaatioOid);
   return (
     <>
       <NavigationAnchor id="hakukohteet" />
@@ -77,6 +99,7 @@ const HakukohteetSection = ({ organisaatioOid, canCreate = true }) => {
         defaultOpen
       >
         <EntitySearchList
+          ActionBar={HakukohdeActionBar}
           searchEntities={searchHakukohteet}
           organisaatioOid={organisaatioOid}
           entityType={HAKUKOHDE}
