+++ src/main/app/src/pages/haku/HakuForm/HakukohteetSection.tsx
@@ -1,61 +1,40 @@
-import React, { useCallback, useMemo } from 'react';
+import React, { useMemo } from 'react';
 
 import { useInterpret } from '@xstate/react';
 import { useTranslation } from 'react-i18next';
 
+import {
+  makeModifiedColumn,
+  makeNimiColumn,
+  makeTilaColumn,
+  makeOrganisaatioColumn,
+} from '#/src/components/ListTable';
 import { ENTITY } from '#/src/constants';
 import { hakukohdeMachine } from '#/src/machines/filterMachines';
-import { createHakukohdeListColumns } from '#/src/pages/haku/HakuForm/createHakukohdeListColumns';
-import { EntityListActionBar } from '#/src/pages/HomePage/EntityListActionBar';
 import { EntitySearchList } from '#/src/pages/HomePage/EntitySearchList';
-import {
-  SERVICE_BY_ENTITY,
-  useEntitySelection,
-} from '#/src/pages/HomePage/useEntitySelection';
 import { useFilterState } from '#/src/pages/HomePage/useFilterState';
 import { searchFilteredHakukohteet } from '#/src/utils/hakukohde/searchHakukohteet';
-import isYhteishakuHakutapa from '#/src/utils/isYhteishakuHakutapa';
 
-const { HAKUKOHDE } = ENTITY;
-
-const HakukohdeActionBar = () => {
-  const { selection, removeSelection } = useEntitySelection(HAKUKOHDE);
-
-  const changeTila = useCallback(
-    tila => {
-      //TODO vaihdetaanko heti vai tallennuksen yhteydessä?
-      console.log('testi tila');
-      console.log(tila);
-      console.log('testi selection');
-      console.log(selection);
-    },
-    [selection]
-  );
-  return (
-    <EntityListActionBar
-      entityType={HAKUKOHDE}
-      selection={selection}
-      removeSelection={removeSelection}
-      changeTila={changeTila}
-      copyEntities={undefined}
-    />
-  );
-};
-
-export const HakukohteetSection = function ({ haku, organisaatioOid }) {
-  const { t } = useTranslation();
-  const isYhteishaku = isYhteishakuHakutapa(haku?.hakutapaKoodiUri);
-
-  const columns = useMemo(
-    () =>
-      createHakukohdeListColumns(
-        t,
-        organisaatioOid
-      )(SERVICE_BY_ENTITY[HAKUKOHDE]),
+const useTableColumns = (t, organisaatioOid) =>
+  useMemo(
+    () => [
+      makeNimiColumn(t, {
+        getLinkUrl: ({ oid }) =>
+          `/organisaatio/${organisaatioOid}/hakukohde/${oid}/muokkaus`,
+      }),
+      makeOrganisaatioColumn(t),
+      makeTilaColumn(t),
+      makeModifiedColumn(t),
+    ],
     [t, organisaatioOid]
   );
 
-  const actionBar = isYhteishaku ? HakukohdeActionBar : undefined;
+const { HAKUKOHDE } = ENTITY;
+
+export const HakukohteetSection = function ({ haku, organisaatioOid }) {
+  const { t } = useTranslation();
+
+  const columns = useTableColumns(t, organisaatioOid);
 
   let filterParams = { hakuOid: haku?.oid };
 
@@ -65,7 +44,6 @@ export const HakukohteetSection = function ({ haku, organisaatioOid }) {
 
   return (
     <EntitySearchList
-      ActionBar={actionBar}
       searchEntities={searchFilteredHakukohteet(filterParams)}
       organisaatioOid={organisaatioOid}
       entityType={HAKUKOHDE}
